feat: implement authentication middleware and update event routes to require user authentication

+++ 10- Course Project Build a REST API incl Authentication SQL Database/models/event.go
@@ -11,11 +11,11 @@ type Event struct {
 	Description string `binding:"required"`
 	Location string `binding:"required"`
 	DateTime time.Time `binding:"required"`
-	UserID int 
+	UserID int64
 
 }
 
-func New(id int64, name string, description string, location string, dateTime time.Time, userID int) Event {
+func New(id int64, name string, description string, location string, dateTime time.Time, userID int64) Event {
 	return Event{
 		ID:          id,
 		Name:       name,
